Refactoring auth
